feat(dream): convergence guard + single-owner lock + exact lookback
Three improvements so re-running dream is cheap and safe:

- Convergence guard: stamp a corpus fingerprint (count + MAX(updated) of meta)
  in the receipt; on the next run, if signal-gather added nothing AND the
  fingerprint is unchanged, skip the expensive LLM passes (contradict /
  synthesize / consolidate). A re-run on an idle corpus is now ~instant instead
  of redoing identical work. --force overrides. Conservative: the nightly still
  runs full whenever there's new transcript activity.
- Single-owner flock (.dream.lock): a second dream run (manual, or the
  com.memo.dream LaunchAgent firing mid-run) now skips instead of racing on the
  shared sidecar DBs and clobbering last.json. Mirrors sync_once's lock.
- Exact signal-gather lookback: use the precise last_run_ts in fractional days
  (+~30min overlap) instead of day-rounded max(1,days)+1, which re-mined ~1-2
  days of transcripts every run. Types widened int->float through the miner.

Verified on an isolated corpus: 2nd run converges (1s, no LLM), --force defeats
it, a new memory flips the fingerprint, and a held lock makes a concurrent run
skip.

+++ src/memo/cli_dream_passes.py
@@ -90,7 +90,7 @@ def _build_orientation(mem: Memory) -> dict:
     return result
 
 
-def _run_signal_gather(since_days: int, file_limit: int = 20) -> dict:
+def _run_signal_gather(since_days: float, file_limit: int = 20) -> dict:
     """Run transcript mining and return a compact summary.
 
     Never raises — exceptions are captured in the returned dict.
